+++ R-Type/shared/resources/melee/melee.hpp
@@ -6,10 +6,10 @@ struct MYLIB_EXPORT Melee : public IResource {
    								    Melee(void) = default;
     virtual 					    ~Melee(void) = default;
     virtual std::string             getName(void) const;
-    virtual double                  getHeight(void) const;
-    virtual double                  getWidth(void) const;
-	virtual int                     getLife(void) const;
-	virtual double 			        getMoveSpeed(void) const;
+    virtual short                  getHeight(void) const;
+    virtual short                  getWidth(void) const;
+    virtual short                     getLife(void) const;
+    virtual double 			        getMoveSpeed(void) const;
     virtual double                  getFireDeltaTime(void) const;
     virtual IResource::Type         getType(void) const;
 	virtual IResource::Vec2         move(double deltaTime) const;
